Separate translations logic
